chore: split exporter tests into test_run_exporter.py for pylint module-size limit
test_run.py crossed pylint's max-module-lines (1043/1000) after the new
notification tests. Move the run.exporter()-level tests (dispatch, shared
tail, return value, stop wiring, and the run() notification tests built on
the same helpers) into tests/unit/test_run_exporter.py -- pure move, no
test changes; 683 tests before and after.

Also simplify an empty-dict equality assert flagged by pylint C1803.
